Lift out common Konsist functionality for reusability in test source set.
