+++ app/src/test/java/com/developersbreach/composeactors/konsist/DesignSystemTest.kt
@@ -1,7 +1,5 @@
 package com.developersbreach.composeactors.konsist
 
-import com.lemonappdev.konsist.api.Konsist
-import com.lemonappdev.konsist.api.verify.assertFalse
 import org.junit.Test
 
 /**
@@ -9,17 +7,13 @@ import org.junit.Test
  * only within the `designSystem` package.
  * These tests checks for direct imports of specified components outside the allowed
  * files, ensuring consistency and maintainability across the UI.
- *
- * Usage:
- * - `materialComponent`: The fully qualified name of the component to check (e.g., [androidx.compose.material.Text]).
- * - `excludePaths`: Paths to files within `designSystem` where direct usage is allowed.
  */
 class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose text should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Text",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Text",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Text.kt",
                 "design-system/src/main/java/com/developersbreach/designsystem/components/TextField.kt",
@@ -29,8 +23,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose textField should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.TextField",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.TextField",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/TextField.kt"
             )
@@ -39,8 +33,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose snackbar should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Snackbar",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Snackbar",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Snackbar.kt"
             )
@@ -49,8 +43,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose surface should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Surface",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Surface",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Surface.kt"
             )
@@ -59,8 +53,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose card should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Card",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Card",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Card.kt"
             )
@@ -69,8 +63,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose image should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.foundation.Image",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.foundation.Image",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Image.kt"
             )
@@ -79,8 +73,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose icon should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Icon",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Icon",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Icon.kt"
             )
@@ -89,8 +83,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose iconButton should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.IconButton",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.IconButton",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/IconButton.kt"
             )
@@ -99,8 +93,8 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose scaffold should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Scaffold",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Scaffold",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Scaffold.kt"
             )
@@ -109,29 +103,11 @@ class DesignSystemTest {
 
     @Test
     fun `no direct usage of androidx compose divider should be allowed except designSystem`() {
-        checkNoDirectUsageExceptDesignSystem(
-            materialComponent = "androidx.compose.material.Divider",
+        checkNoDirectUsageExceptAllowed(
+            componentName = "androidx.compose.material.Divider",
             excludePaths = arrayOf(
                 "design-system/src/main/java/com/developersbreach/designsystem/components/Divider.kt"
             )
         )
-    }
-
-    private fun checkNoDirectUsageExceptDesignSystem(
-        materialComponent: String,
-        excludePaths: Array<String>
-    ) {
-        Konsist.scopeFromProject()
-            .files
-            .filter { file ->
-                val normalizedFilePath = file.path.replace("\\", "/").lowercase()
-                excludePaths.none { normalizedFilePath.contains(it.replace("\\", "/").lowercase()) }
-            }
-            .flatMap { file ->
-                file.imports.filter { it.name == materialComponent }.map { file }
-            }
-            .assertFalse {
-                true // Ensure that the Material component is not directly used outside the designSystem
-            }
     }
 }
