--- app/src/test/java/com/developersbreach/composeactors/konsist/PreviewTest.kt
@@ -1,28 +1,0 @@
-package com.developersbreach.composeactors.konsist
-
-import com.lemonappdev.konsist.api.Konsist
-import com.lemonappdev.konsist.api.verify.assertFalse
-import org.junit.Test
-
-class PreviewTest {
-
-    @Test
-    fun `no direct usage of androidx compose Preview should be allowed except PreviewLightDark`() {
-        val previewAnnotation = "androidx.compose.ui.tooling.preview.Preview"
-        val excludePaths = arrayOf(
-                "compose-actors/app/src/main/java/com/developersbreach/composeactors/annotations/PreviewLightDark.kt"
-            )
-        Konsist.scopeFromProject()
-            .files
-            .filter { file ->
-                val normalizedFilePath = file.path.replace("\\", "/").lowercase()
-                excludePaths.none { normalizedFilePath.contains(it.replace("\\", "/").lowercase()) }
-            }
-            .flatMap { file ->
-                file.imports.filter { it.name == previewAnnotation }.map { file }
-            }
-            .assertFalse {
-                true
-            }
-    }
-}
